Update the new SDK teleport example

+++ examples/new-sdk/teleport/terminate.sh
@@ -1,14 +1,16 @@
 #!/usr/bin/env sh
-TEST_PIDS=$(ps aux | grep python | grep -E "teleport" | awk {'print $2'})
+TEST_PIDS=$(ps aux | grep python | grep -E "Test" | awk {'print $2'})
 if [ "$TEST_PIDS" != "" ]
 then
         kill -9 $TEST_PIDS
 fi
 
-simulaqron stop
-
 # Check if SimulaQron is running
 if [ -f ~/.simulaqron_pids/simulaqron_network_default.pid ]; then
-    cat $HOME/.simulaqron_pids/simulaqron_network_default.pid | xargs kill -9 
+    if ! simulaqron stop
+    then
+        # Kill the process, only if simulaqron could not be stopped gracefully
+        cat $HOME/.simulaqron_pids/simulaqron_network_default.pid | xargs kill -9
+    fi
 fi
